Refactoring: Made it a class + minor cleanup.
Okay, almost the final step. Now there's a class "Explainer" that does
all the work. It offers a method called "explain(text)" that explains
the given text. Options can be set as member variables.

As a result, we get a nice and clean API:

	import explain

	mystr = """... explain-syntax ..."""

	explainer = explain.Explainer()
	explainer.straight = '!'                # optional, of course
	expl_list = explainer.explain(mystr)
	print '\n'.join(expl_list),

As a bonus, it's now easier to avoid all that "adding trailing
newlines"-foo.