Refactoring: Reverse indexed_comments when it's needed.
Only annotate_indexed_comments() needs indexed_comments to be reversed.
Ergo, don't reverse it in parse_plaintext_explanation().

Again, this results in a cleaner API. You really don't expect to get a
reversed list after a call to parse_plaintext_explanation(). Reversing
the list is just a detail of the implementation of
annotate_indexed_comments() -- it should not be visible to the outside.