Added support for manual line wrapping.
Basically, a user wanted to do something like this:

	Number 1
	       |
	       \- 1: One!
	          2: Two!
	          3: Three!

As that comment is very short, the old code has automatically wrapped
it. So, it came out like this:

	Number 1
	       |
	       \- 1: One! 2: Two! 3: Three!

That's not what he was after. His manual line breaks were missing. (Of
course they were, because the script kicked out any newlines and
re-wrapped the text.)

This commit introduces a new syntax item to denote explicit line breaks.
You can now end a line in a comment with "\\" (two backslashes). Doing
so inserts a line break at the end of this line which will *not* be
removed.

Here's an example:

	Number 1
	------ -

	This is a very long line. There's a lot of text. It'll get wrapped
	automatically. Also note that there's line breaks inside of this
	comment. They'll be removed. This is the "traditional" way of handling
	comments.

	1: One! \\
	2: Two! \\
	3: Three! \\
	Note: I added '\\' at the ends of those lines.

And that's the result:

	Number 1
	\____/ |
	   |   \- 1: One!
	   |      2: Two!
	   |      3: Three!
	   |      Note: I added '\\' at the ends of those lines.
	   |
	   \- This is a very long line. There's a lot of text. It'll get wrapped
	      automatically. Also note that there's line breaks inside of this
	      comment. They'll be removed. This is the "traditional" way of
	      handling comments.

Now, this patch is pretty short. I admit that the way it works may not
be the best solution. If you come up with something better (without
having to rewrite the whole script), please tell me! :)

Here's why this works:

Usually, all newlines get stripped from comments -- directly after a
line has been read. So, after pre-processing a comment, it is guaranteed
not to have any newlines in it. This means: We can artificially insert
new newlines in order to denote manual line breaks *after*
pre-processing.

Hence, a trailing '\\' gets replaced with a newline.

Later on, comments are split up again according to newlines. This honors
manually inserted newlines. Then, each line will be wrapped
individually.